+++ lib/applitools/connectivity/server_connector.rb
@@ -284,7 +284,6 @@ module Applitools::Connectivity
         endpoint_url, body: request_body
       )
       raise Applitools::EyesError.new("Request failed: #{res.status} #{res.body} #{request_body}") unless res.success?
-
       response = Oj.load(res.body)
       Applitools::Session.new(response, res.status == HTTP_STATUS_CODES[:created])
     end