+++ lib/applitools/selenium/visual_grid/i_render_browser_info.rb
@@ -7,18 +7,9 @@ module Applitools
       string_field :platform
       string_field :size_mode
       string_field :baseline_env_name
-      object_field :emulation_info, Applitools::Selenium::EmulationBaseInfo
-      object_field :ios_device_info, Applitools::Selenium::EmulationBaseInfo
-
-      def device_name
-        return 'desktop' unless emulation_info || ios_device_info
-        return ios_device_info.device_name if ios_device_info
-        return emulation_info.device_name + ' (chrome emulation)' if emulation_info
-      end
 
       def to_s
-        return "#{viewport_size} (#{browser_type})" unless emulation_info
-        "#{emulation_info.device_name} - #{emulation_info.screen_orientation}"
+        return "#{viewport_size} (#{browser_type})"
       end
     end
   end