+++ spec/integration/classic_api_spec.rb
@@ -1,6 +1,10 @@
 # frozen_string_literal: true
 require_relative 'test_classic_api_v1'
-RSpec.describe 'Classic API' do
+RSpec.describe 'Classic API', report_me: true do
+  def test_reporting_group
+    'selenium'
+  end
+
   context 'Eyes Selenium SDK - Classic API', selenium: true do
     include_examples 'Classic API'
   end