+++ spec/integration/duplicates_spec.rb
@@ -1,6 +1,10 @@
 # frozen_string_literal: true
 require_relative 'test_duplicates_v1'
-RSpec.describe 'duplicates' do
+RSpec.describe 'duplicates', report_me: true do
+  def test_reporting_group
+    'selenium'
+  end
+
   context 'Eyes Selenium SDK - Duplicates', selenium: true do
     include_examples 'Eyes Selenium SDK - Duplicates'
   end