+++ spec/integration/fluent_api_spec.rb
@@ -1,6 +1,10 @@
 # frozen_string_literal: true
 require_relative 'test_fluent_api_v1'
-RSpec.describe 'Fluent API' do
+RSpec.describe 'Fluent API', report_me: true do
+  def test_reporting_group
+    'selenium'
+  end
+
   context 'Eyes Selenium SDK - Fluent API', selenium: true do
     include_examples 'Fluent API'
   end