+++ spec/integration/simple_cases_spec.rb
@@ -1,7 +1,10 @@
 # frozen_string_literal: true
 require_relative 'test_simple_cases_v1'
+RSpec.describe 'Simple test cases', report_me: true do
+  def test_reporting_group
+    'selenium'
+  end
 
-RSpec.describe 'Simple test cases' do
   context 'Eyes Selenium SDK - Simple Test Cases', selenium: true do
     include_examples 'Eyes Selenium SDK - Simple Test Cases'
   end